+++ app/src/main/java/com/example/moviedb/ui/screen/MainActivity.kt
@@ -1,14 +1,8 @@
 package com.example.moviedb.ui.screen
 
-import android.os.Bundle
-import androidx.navigation.findNavController
-import androidx.navigation.plusAssign
-import androidx.navigation.ui.setupWithNavController
 import com.example.moviedb.R
 import com.example.moviedb.databinding.ActivityMainBinding
 import com.example.moviedb.ui.base.BaseActivity
-import com.example.moviedb.ui.navigation.KeepStateNavigator
-import kotlinx.android.synthetic.main.activity_main.*
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class MainActivity : BaseActivity<ActivityMainBinding, MainActivityViewModel>() {
@@ -17,25 +11,4 @@ class MainActivity : BaseActivity<ActivityMainBinding, MainActivityViewModel>() 
 
     override val layoutId: Int = R.layout.activity_main
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        val navController = findNavController(R.id.nav_host_fragment)
-
-        // get fragment
-        val navHostFragment = supportFragmentManager.findFragmentById(R.id.nav_host_fragment)!!
-
-        // setup custom navigator
-        val navigator = KeepStateNavigator(
-            context = this,
-            manager = navHostFragment.childFragmentManager,
-            containerId = R.id.nav_host_fragment
-        )
-        navController.navigatorProvider += navigator
-
-        // set navigation graph
-        navController.setGraph(R.navigation.nav_graph)
-
-        bottom_nav?.setupWithNavController(navController)
-    }
 }
