+++ app/src/main/java/com/example/moviedb/ui/screen/main/MainViewModel.kt
@@ -1,10 +1,6 @@
 package com.example.moviedb.ui.screen.main
 
-import androidx.lifecycle.MutableLiveData
 import com.example.moviedb.ui.base.BaseViewModel
 
 class MainViewModel : BaseViewModel() {
-
-    val currentTab = MutableLiveData<Int>()
-
 }
