+++ app/src/main/java/com/example/moviedb/ui/screen/oldmain/OldMainViewModel.kt
@@ -1,0 +1,10 @@
+package com.example.moviedb.ui.screen.oldmain
+
+import androidx.lifecycle.MutableLiveData
+import com.example.moviedb.ui.base.BaseViewModel
+
+class OldMainViewModel : BaseViewModel() {
+
+    val currentTab = MutableLiveData<Int>()
+
+}
