+++ app/src/main/java/com/example/moviedb/ui/screen/splash/SplashViewModel.kt
@@ -1,0 +1,6 @@
+package com.example.moviedb.ui.screen.splash
+
+import com.example.moviedb.ui.base.BaseViewModel
+
+class SplashViewModel : BaseViewModel() {
+}
