+++ app/src/main/res/navigation/nav_graph_main.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_graph_main"
+    app:startDestination="@id/navigation_popular">
+
+    <keep_state_fragment
+        android:id="@+id/navigation_popular"
+        android:name="com.example.moviedb.ui.navigation.popularcontainer.PopularContainerFragment"
+        android:label="PopularContainerFragment"
+        tools:layout="@layout/fragment_popular_container" />
+
+    <keep_state_fragment
+        android:id="@+id/navigation_favorite"
+        android:name="com.example.moviedb.ui.navigation.favoritecontainer.FavoriteContainerFragment"
+        android:label="FavoriteContainerFragment"
+        tools:layout="@layout/fragment_favorite_container" />
+
+    <keep_state_fragment
+        android:id="@+id/navigation_tv"
+        android:name="com.example.moviedb.ui.screen.tv.TvListFragment"
+        android:label="TvListFragment"
+        tools:layout="@layout/fragment_loadmore_refresh" />
+
+</navigation>
